backend/requirements.txt
Django==4.2.1
psycopg2-binary==2.9.3
coveralls==3.3.1
Faker==18.7.0
djangorestframework>=3.12.4
drf_spectacular==0.26.1
gunicorn==20.1.0

# display tracebacks correctly in parallel test runner
tblib==1.7.0

# code coverage
coverage==6.2
coveralls==3.3.1
flake8==6.1.0
pandas==2.0.3
